@@ -9,8 +9,9 @@ import (
 	"github.com/jedib0t/go-pretty/v6/table"
 	"github.com/spf13/cobra"
 	"github.com/whitekid/goxp/log"
-	"github.com/whitekid/reader"
-	"github.com/whitekid/reader/db"
+
+	"reader"
+	"reader/db"
 )
 
 func init() {
