@@ -3,13 +3,11 @@ package db
 import (
 	"context"
 	"os"
-	"path"
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	"github.com/whitekid/goxp"
-	"github.com/whitekid/goxp/fixtures"
-	"github.com/whitekid/reader/testutils"
+
+	"reader/testutils"
 )
 
 func must(err error) {
@@ -21,8 +19,6 @@ func must(err error) {
 func TestMain(m *testing.M) {
 	sqlDB, err := Open("test.db")
 	must(err)
-
-	fixtures.Database(sqlDB, "sqlite", path.Join(path.Dir(goxp.Filename()), "../fixtures/testdata"))
 
 	testutils.SetupFixtureDatabase(sqlDB)
 	os.Exit(m.Run())
